+++ test/VisualStudio.Test/ProjectExtensionsTest.cs
@@ -1,11 +1,12 @@
-﻿using EnvDTE;
-using Xunit;
+﻿using System.IO;
+using EnvDTE;
 using Moq;
+using Xunit;
+using Xunit.Extensions;
 
 namespace NuGet.VisualStudio.Test
 {
     using PackageUtility = NuGet.Test.PackageUtility;
-    using System.IO;
 
     public class ProjectExtensionsTest
     {
@@ -20,6 +21,100 @@ namespace NuGet.VisualStudio.Test
 
             // Assert
             Assert.Equal(@"WebProject\Bin", path);
+        }
+
+        [Fact]
+        public void GetTargetFrameworkForXnaProjectReturnsWindowsPhoneTargetFramework()
+        {
+            // Arrange
+            var project = new Mock<Project>();
+
+            var xnaProperty = new Mock<Property>();
+            xnaProperty.Setup(x => x.Value).Returns("Windows Phone OS 7.1");
+
+            project.Setup(p => p.Properties.Item(It.Is<object>(v => "Microsoft.Xna.GameStudio.CodeProject.WindowsPhoneProjectPropertiesExtender.XnaRefreshLevel".Equals(v))))
+                   .Returns(xnaProperty.Object);
+
+            var fxProperty = new Mock<Property>();
+            fxProperty.Setup(x => x.Value).Returns(".NETFramework,Version=v4.0");
+            project.Setup(p => p.Properties.Item(It.Is<object>(v => "TargetFrameworkMoniker".Equals(v))))
+                   .Returns(fxProperty.Object);
+
+            // Act
+            string targetFramework = ProjectExtensions.GetTargetFramework(project.Object);
+
+            // Assert
+            Assert.Equal("Silverlight,Version=v4.0,Profile=WindowsPhone71", targetFramework);
+        }
+
+        [Fact]
+        public void GetTargetFrameworkForWrongXnaProjectDoesNotReturnWindowsPhoneTargetFramework()
+        {
+            // Arrange
+            var project = new Mock<Project>();
+
+            var xnaProperty = new Mock<Property>();
+            xnaProperty.Setup(x => x.Value).Returns("Windows Phone OS 7.0");    // 7.0 is not recognized. Only 7.1 is.
+
+            project.Setup(p => p.Properties.Item(It.Is<object>(v => "Microsoft.Xna.GameStudio.CodeProject.WindowsPhoneProjectPropertiesExtender.XnaRefreshLevel".Equals(v))))
+                   .Returns(xnaProperty.Object);
+
+            var fxProperty = new Mock<Property>();
+            fxProperty.Setup(x => x.Value).Returns(".NETFramework,Version=v4.0");
+            project.Setup(p => p.Properties.Item(It.Is<object>(v => "TargetFrameworkMoniker".Equals(v))))
+                   .Returns(fxProperty.Object);
+
+            // Act
+            string targetFramework = ProjectExtensions.GetTargetFramework(project.Object);
+
+            // Assert
+            Assert.Equal(".NETFramework,Version=v4.0", targetFramework);
+        }
+
+        [Fact]
+        public void GetTargetFrameworkForMissingXnaPropertyDoesNotReturnWindowsPhoneTargetFramework()
+        {
+            // Arrange
+            var project = new Mock<Project>();
+
+            var fxProperty = new Mock<Property>();
+            fxProperty.Setup(x => x.Value).Returns(".NETFramework,Version=v4.0");
+            project.Setup(p => p.Properties.Item(It.Is<object>(v => "TargetFrameworkMoniker".Equals(v))))
+                   .Returns(fxProperty.Object);
+
+            // Act
+            string targetFramework = ProjectExtensions.GetTargetFramework(project.Object);
+
+            // Assert
+            Assert.Equal(".NETFramework,Version=v4.0", targetFramework);
+        }
+
+        [Theory]
+        [InlineData(".NETFramework,Version=v4.5")]
+        [InlineData("Silverlight,Version=v4.0")]
+        [InlineData("WindowsPhone,Version=v7.0")]
+        [InlineData("WindowsPhone,Version=v8.0")]
+        public void GetTargetFrameworkForWrongTargetFrameowrkVersionInXnaProjectDoesNotReturnWindowsPhoneTargetFramework(string framework)
+        {
+            // Arrange
+            var project = new Mock<Project>();
+
+            var xnaProperty = new Mock<Property>();
+            xnaProperty.Setup(x => x.Value).Returns("Windows Phone OS 7.1");
+
+            project.Setup(p => p.Properties.Item(It.Is<object>(v => "Microsoft.Xna.GameStudio.CodeProject.WindowsPhoneProjectPropertiesExtender.XnaRefreshLevel".Equals(v))))
+                   .Returns(xnaProperty.Object);
+
+            var fxProperty = new Mock<Property>();
+            fxProperty.Setup(x => x.Value).Returns(framework);
+            project.Setup(p => p.Properties.Item(It.Is<object>(v => "TargetFrameworkMoniker".Equals(v))))
+                   .Returns(fxProperty.Object);
+
+            // Act
+            string targetFramework = ProjectExtensions.GetTargetFramework(project.Object);
+
+            // Assert
+            Assert.Equal(framework, targetFramework);
         }
 
         [Fact]